refactor: move top-level from_* factories into ada/factories.py
from_ifc / from_step / from_acis / from_fem / from_fem_res / from_sesam_cc / from_genie_xml move
out of ada/__init__.py into ada/factories.py and are re-exported, so ada.from_step(...) is
unchanged. Keeps the package __init__ lean (drops ~165 lines + now-unused os/pathlib/TYPE_CHECKING
imports). Assembly-returning factories gain an optional cad_config attached to the result.
